API/requirements.txt
Flask
gunicorn
numpy==1.19.5
Pillow==8.1.0
imageio==2.9.0
Keras==2.4.3
pyyaml==5.3.1
scipy==1.6.0
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
keras-preprocessing==1.1.2
markdown==3.3.3
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
wheel==0.36.2
wrapt==1.12.1
opencv-python==4.5.1.48
scikit-learn==0.24.1